refactor: move colors.js from legacy root to legacy/lib directory
- Move colors.js from /legacy to /legacy/lib for better organization
- Update import paths in legacy/bin/nullvoid.js and legacy/lib/logger.js
- Maintains same functionality while improving code structure
