+++ legacy/lib/colors.js
